+++ src/client/index.ts
@@ -1,4 +1,5 @@
 export { SoundCloudClient } from "./SoundCloudClient.js";
 export type { SoundCloudClientConfig, TokenOption } from "./SoundCloudClient.js";
-export { scFetch } from "./http.js";
+export { scFetch, scFetchUrl } from "./http.js";
 export type { RequestOptions, AutoRefreshContext } from "./http.js";
+export { paginate, paginateItems, fetchAll } from "./paginate.js";